% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_covid_data.R
\docType{data}
\name{aus_covid_data}
\alias{aus_covid_data}
\title{Daily COVID19 cases and deaths in Australia by state}
\format{
An object of class \code{data.frame} with 6552 rows and 7 columns.
}
\source{
The raw data pulled from the Johns Hopkins University Center for Systems Science and Engineering (JHU CCSE) Coronavirus repository.
}
\usage{
aus_covid_data
}
\description{
This data set contains COVID19 related data from states and territories in Australia from the 22nd of January to the most recent date. There are 7 variables in the dataset including province representing states and territories in Australia, type of cases including confirmed, death and recovered.
}
\keyword{datasets}
